requirements.txt
pyaudio
webrtcvad
setuptools
numpy
google-generativeai
